Rework completecases and nonunique
Several people have noted that they expect completecases to return a
datatable of the actual complete cases, so even though we reworked the
dropnull functions to try and abate this issue, it doesn't appear that
we chose the best fix. Several people have noted that nonunique is also
confusing. This implements the suggestions for addressing these issues
presented in
https://github.com/JuliaData/DataTables.jl/issues/38#issuecomment-288556269.
It also includes a fix for the printtable test in introduced which has a
random variable in the test that breaks the test on a regular basis.
I'll open another PR with just this change though.