Extend nullable! code, fix NullableArray hardcoding
Adds nullable versions of nullable! for copies rather than inplace.
Adds tests to check proper handling of CategoricalArrays. Adds a
similar_nullable function that will use the size of the Array rather
than always requiring the user to pass the size as an argument. Converts
a nullable! call that operated over the entire datatable to use a new
nullable! func that operates over all columns.