Fix printing CategoricalValue
CategoricalValue entries should always be printed via showcompact() in order
to get a short representation. This uses ourshowcompact() to do that when
printing DataTables to REPL via show, as well as when printing them to HTML,
LaTeX and CSV via printtable().

Also fix a failure due to duplicated keyword arguments on Julia 0.7.